Potential fix for double frontmatter injection
Potential fix for double frontmatter injection if frontmatterText couldn't be extracted successfully.
Tested locally and fixed the problem for me. Seems to be related to frontmatter
not being extracted successfully or the eventId being written within injectEventDetails and thus
not being present when the eventId was injected first. This is the reason for the call order change.